+++ src/RuleEvaluators/TimeEvaluator.php
@@ -3,7 +3,6 @@
 namespace JobMetric\Rolix\RuleEvaluators;
 
 use Carbon\Carbon;
-use DateTimeZone;
 use JobMetric\CustomField\CustomFieldBuilder;
 use JobMetric\Form\FormBuilder;
 use JobMetric\Rolix\Contracts\AbstractRuleEvaluator;
@@ -69,7 +68,7 @@ class TimeEvaluator extends AbstractRuleEvaluator
                     ->label('rolix::base.rule_evaluators.time.timezone.label')
                     ->info('rolix::base.rule_evaluators.time.timezone.info')
                     ->validation('nullable|string|timezone')
-                    ->options(array_map(static fn (string $timezone) => ['label' => $timezone, 'value' => $timezone], DateTimeZone::listIdentifiers()));
+                    ->options($this->timezoneOptions());
             });
         });
     }